Refactored user services, added controllers and services for auxiliary data, and improved error handling

+++ ScooterRent/src/main/java/com/example/repository/RentalPointRepository.java
@@ -3,5 +3,5 @@ package com.example.repository;
 import com.example.entity.RentalPoint;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RentalPointRepository extends JpaRepository<RentalPoint, Integer> {
+public interface RentalPointRepository extends JpaRepository<RentalPoint, Long> {
 }
